requirements.txt
Flask==2.2.5
Flask_Cors==4.0.1
numpy==1.24.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==10.4.0



# To auto generate main requirements run the following
# pip install pipreqs
# pipreqs . --force